Proof-of-concept high-speed driver
The serial example builds, enumerates, and serves its simple purpose.
The test class builds as well. When I run the test suite on my host,

  cargo test --features test-class-high-speed

I hit a panic:

  [ERROR support]: panicked at 'bulk read InvalidState', /Users/mciantyre/.cargo/git/checkouts/usb-device-e37115d9692a39c7/982a988/src/test_class.rs:140:25

So we're probably seeing a high flag in the TD status bits. Next step
is to figure out which bit, and figure out how to recover. The manual
has an error matrix that we can consider.